removed deprecation warnings to keep javac quiet
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@949434 13f79535-47bb-0310-9956-ffa450edef68

+++ src/examples/src/org/apache/poi/hssf/usermodel/examples/CellComments.java
@@ -89,7 +89,7 @@ public class CellComments {
          * It works, the comment is visible.
          */
         comment2.setRow(6);
-        comment2.setColumn((short)1);
+        comment2.setColumn(1);
 
         FileOutputStream out = new FileOutputStream("poi_comment.xls");
         wb.write(out);
